app_orchestrator/requirements.txt
celery  == 5.4.0

python-dotenv == 1.1.0

loguru == 0.7.3

pydantic == 2.10.6

pymongo == 4.11.3

loguru == 0.7.3

pika == 1.3.2

requests == 2.32.3